fix: dep rules
