+++ src/store/transactions/fees.hooks.ts
@@ -1,4 +1,4 @@
-import { useAtomValue } from 'jotai/utils';
+import { useAtomCallback, useAtomValue } from 'jotai/utils';
 import {
   currentFeeRateState,
   currentFeeState,
@@ -8,6 +8,10 @@ import {
   feeRateUseCustom,
 } from '@store/transactions/fees';
 import { useAtom } from 'jotai';
+import { useRawTxIdState } from '@store/transactions/raw.hooks';
+import { useSubmitTransactionCallback } from '@pages/transaction-signing/hooks/use-submit-stx-transaction';
+import { useCallback } from 'react';
+import { rawSignedStacksTransactionState } from '@store/transactions/raw';
 
 export function useCurrentFee() {
   return useAtomValue(currentFeeState);
@@ -32,3 +36,32 @@ export function useFeeRate() {
 export function useCurrentFeeRate() {
   return useAtom(currentFeeRateState);
 }
+
+export const useReplaceByFeeSubmitCallBack = () => {
+  const [, setMultiplier] = useFeeRateMultiplierCustom();
+  const [, setUseCustom] = useFeeRateUseCustom();
+  const [, setFeeRate] = useFeeRate();
+  const [, setTxId] = useRawTxIdState();
+
+  const submitTransaction = useSubmitTransactionCallback({
+    onClose: () => {
+      setTxId(null);
+      setUseCustom(false);
+      setMultiplier(undefined);
+      setFeeRate(undefined);
+    },
+    loadingKey: 'speed',
+    replaceByFee: true,
+  });
+
+  return useAtomCallback<void, { fee: number; nonce: number }>(
+    useCallback(
+      async get => {
+        const signedTx = await get(rawSignedStacksTransactionState, true);
+        if (!signedTx) return;
+        await submitTransaction(signedTx);
+      },
+      [submitTransaction]
+    )
+  );
+};
